requirements.txt
beautifulsoup4==4.11.1
fpdf2==2.8.1
arxiv==2.1.3
openai==1.54.3
pandas==2.2.3
rapidfuzz==3.10.1
pyarrow==18.0.0
fastparquet==2024.5.0